Big refactor to a more sensible package structure
And fix everything up.
