+++ server/trillian_log_signer/main.go
@@ -28,8 +28,8 @@ import (
 	"github.com/google/trillian/monitoring/metric"
 	"github.com/google/trillian/quota"
 	"github.com/google/trillian/server"
-	"github.com/google/trillian/storage/coresql"
-	"github.com/google/trillian/storage/db"
+	"github.com/google/trillian/storage/sql/coresql"
+	"github.com/google/trillian/storage/sql/coresql/db"
 	"github.com/google/trillian/util"
 	"github.com/google/trillian/util/etcd"
 	"golang.org/x/net/context"
