+++ storage/mysql/admin_storage_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/google/trillian"
 	"github.com/google/trillian/crypto/keyspb"
 	"github.com/google/trillian/storage"
-	"github.com/google/trillian/storage/coresql"
+	"github.com/google/trillian/storage/sql/coresql"
 	"github.com/google/trillian/storage/testonly"
 )
