+++ storage/tools/fetch_leaves/main.go
@@ -22,8 +22,8 @@ import (
 	_ "github.com/go-sql-driver/mysql" // Load MySQL driver
 
 	log "github.com/golang/glog"
-	"github.com/google/trillian/storage/coresql"
-	"github.com/google/trillian/storage/db"
+	"github.com/google/trillian/storage/sql/coresql"
+	"github.com/google/trillian/storage/sql/coresql/db"
 )
 
 var (
